requirements.txt
langchain>=0.1.12
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-zhipuai>=0.1.0
python-dotenv==1.0.1
zhipuai==1.0.7